Updated to django-q 1.3.2
I took the changes made here for cluster.py and merged them with the master version of @Koed00/django-q/cluster.py for version 1.3.2.

I've tried running it (Windows 10 with an AMD Ryzen7 [not sure if that matters], Django 3.0.6, django-q 1.3.2 with this file, using sqlite db as backend for both site and django-q):

If I set 'sync': True in settings.py, I get an error 'NoneType' has no attribute 'increment' every 30 seconds or so, and it ignores tasks in the Scheduled Tasks table. However, if i let it default to false, I get no error, and I can put a scheduled task in via admin. That task gets enqueued, but I haven't actually seen django-q try to execute it yet (no errors in the console, and no entries in successful tasks or failed tasks).

Thought this might be helpful; it'd be awesome to see your work incorporated into django-q.